pyproject.toml
[project]
name = "docs_ocr"
version = "0.1.0"
description = "MVP pipeline for OCR-based FIO extraction from document images"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy",
    "pillow",
    "paddleocr>=3.4.1",
    "paddlepaddle==3.2.2",
]

[dependency-groups]
test = [
    "pytest>=9.0.3",
]
research = [
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = false
